Refactoring hydra_agent
