+++ hydra_agent/utils/__init__.py
@@ -1,0 +1,1 @@
+from utils.help import help
